move get_display from utils to helpers
